Isolate block production interfaces
