+++ blockprod/src/interface/mod.rs
@@ -13,27 +13,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use common::chain::{Block, Destination, SignedTransaction};
-
-use crate::BlockProductionError;
-
-#[async_trait::async_trait]
-pub trait BlockProductionInterface: Send {
-    /// When called, the Block Builder will start creating blocks at the next tip in chainstate
-    fn start(&self) -> Result<(), BlockProductionError>;
-
-    /// When called, the Block Builder will cancel all current attempts to create blocks
-    /// and won't attempt to do it again for new tips in chainstate or mempool
-    /// Call start() to enable again
-    fn stop(&self) -> Result<(), BlockProductionError>;
-
-    /// Check if the Block Builder is currently connected
-    fn is_connected(&self) -> bool;
-
-    /// Generate a block with the given transactions to the specified reward destination
-    async fn generate_block(
-        &self,
-        reward_destination: Destination,
-        transactions: Vec<SignedTransaction>,
-    ) -> Result<Block, BlockProductionError>;
-}
+pub mod blockprod_interface;
+pub mod blockprod_interface_impl;
